Rename application to uploader
